Refactored target information into TargetInfo struct
Instead of manually comparing information from the triplets there's a
TargetInfo structure which is passed along with a number of methods
allowing to easily reason about the target without having to deal with
triplets all over the map. Instead of creating TargetInfo from scratch
all over the map it's passed along all the way from main so it's much
easier to test the code or implement funny stuff in the future. This
makes access to target() superfluous and thus it's private now. No
optimisation has been done to the accessor methods yet, just the
previous code moved here.

Signed-off-by: Daniel Egger <daniel@eggers-club.de>